requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
chainlit
pymupdf
faiss-cpu
sentence-transformers
google-genai